% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/discretize.R
\name{conTocat}
\alias{conTocat}
\title{Convert continuous values to categorical/discrete values}
\usage{
conTocat(covariates, var_names)
}
\arguments{
\item{covariates}{samples x covariates matrix containing the continous covariates which need to be discretized}

\item{varnames}{the continuous variables that need to be discretized}
}
\description{
This function calculates the quantiles for the data  and assigns a quantile group to every data point
}
\note{
the function returns a modified  samples x covariates matrix with the categorized variables appended to
the end as another column
}
\examples{
conTocat(covariate_matrix,"BMI")
}
\author{
Karthikeyan Murugesan<karthikm at gatech.edu>
}
\keyword{categorize,}
\keyword{discretize}
\keyword{quantile,}

